Document Categories are now Pages

+++ app/Controller/PagesController.php
@@ -2,48 +2,56 @@
 class PagesController extends AppController {
 	public $helpers = array('Html', 'Form', 'Markdown.Markdown');
 	public $components = array('Session');
-	
-	var $uses = array('Page', 'Document', 'CmsUser');
-	
+
+	var $uses = array('Document', 'Page', 'CmsUser');
+
 	public function index() {
-		$this->authenticate();
-		$this->set('title', 'Listing All Pages');
-		$this->set('pages', $this->Page->getPages());
+		$this->set('title', 'Pages');
+		$categories = $this->Page->getCategories();
+		$this->set('categories', $categories);
 	}
-	
-	public function view($id = null) {
+
+	public function view($id) {
 		$this->Page->id = $id;
-		$page = $this->Page->read();
-		$this->set('title', $page['Page']['name']);
-		$this->set('page', $page);
+		$category = $this->Page->read();
+		$documents = $this->Document->getFromCategory($id);
+
+		$title = $category['Page']['name'];
+		$this->set('title', $title);
+		$this->set('category', $category);
+		$this->set('documents', $documents);
 	}
-	
+
 	public function add() {
-		$this->authenticate();
-		$this->set('title', 'Add New Page');
+		$Authentication = new Authentication;
+		$cmsuser = $this->CmsUser->findByUser($Authentication->Username());
+		if (!isset($cmsuser['CmsUser'])) {
+			$this->redirect(array('controller' => 'CmsUsers', 'action' => 'accessdenied'));
+		}
+		$this->set('title', 'Add New Document Category');
 		if ($this->request->is('post')) {
-			if ($this->Page->save($this->request->data)) {
+			if($this->Page->save($this->request->data)) {
 				$this->Session->setFlash('
 					<div class="alert alert-success">
 						<button class="close" data-dismiss="alert">&times;</button>
-						Your page has been added successfully.
+						Your category has been added successfully.
 					</div>
 				');
-				$this->redirect(array('action' => 'view', $this->Post->id));
+				$id = $this->Page->id;
+				$this->redirect(array('action' => 'view', $id));
 			} else {
 				$this->Session->setFlash('
 					<div class="alert alert-error">
 						<button class="close" data-dismiss="alert">&times;</button>
-						Unable to add your page. Please make sure you have filled out all fields.
+						Unable to add your category. Please make sure that you have filled out all fields correctly.
 					</div>
 				');
 			}
 		}
 	}
-	
+
 	public function edit($id) {
-		$this->authenticate();
-		$this->set('title', 'Edit Page');
+		$this->set('title', 'Edit Category');
 		$this->Page->id = $id;
 		$this->set('id', $id);
 		if ($this->request->is('get')) {
@@ -53,7 +61,7 @@ class PagesController extends AppController {
 				$this->Session->setFlash('
 					<div class="alert alert-success">
 						<button class="close" data-dismiss="alert">&times;</button>
-						Page updated successfully.
+						Document updated successfully.
 					</div>
 				');
 				$this->redirect(array('action' => 'view', $id));
@@ -61,41 +69,56 @@ class PagesController extends AppController {
 				$this->Session->setFlash('
 					<div class="alert alert-error">
 						<button class="close" data-dismiss="alert">&times;</button>
-						Unable to update page. Please make sure that you have filled out all fields correctly.
+						Unable to update your category. Please make sure that you have filled out all fields correctly.
 					</div>
 				');
 			}
 		}
 	}
-	
+
 	public function delete($id) {
-		$this->authenticate();
 		if ($this->request->is('get')) {
 			throw new MethodNotAllowedException();
 		}
-		if ($this->Page->delete($id)) {
+		$documentCount = $this->Document->countFromCategory($id);
+		if ($documentCount == 0) {
+			if ($this->Page->delete($id)) {
+				$this->Session->setFlash('
+					<div class="alert alert-success">
+						<button class="close" data-dismiss="alert">&times;</button>
+						Document category deleted successfully.
+					</div>
+				');
+				$this->redirect(array('action' => 'index'));
+			}
+		} else {
+			if ($documentCount == 1) {
+				$text = '
+					<p>
+						The category still contains ' . $documentCount . ' document.
+					</p>
+					<p>
+						You must delete it before attempting to delete the category.
+					</p>
+				';
+			} else {
+				$text = '
+					<p>
+						The category still contains ' . $documentCount . ' documents.
+					</p>
+					<p>
+						You must delete them before attempting to delete the category.
+					</p>
+				';
+			}
+
 			$this->Session->setFlash('
-				<div class="alert alert-success">
+				<div class="alert alert-danger">
 					<button class="close" data-dismiss="alert">&times;</button>
-					Page deleted successfully.
+					' . $text . '
 				</div>
 			');
 			$this->redirect(array('action' => 'index'));
 		}
-	}
-
-	// Authentication Magic
-	function authenticate() {
-		$Authentication = new Authentication;
-		$cmsuser = $this->CmsUser->findByUser($Authentication->Username());
-		if (!isset($cmsuser['CmsUser'])) {
-			$this->redirect(
-				array(
-					'controller' => 'CmsUsers',
-					'action' => 'accessdenied'
-				)
-			);
-		}
-		$this->set('cmsuser', $cmsuser);
 	}
 }
